Tensorboard (#36)
* Fix Spelling

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>

* Log cost to TensorBoard

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>

* Fix typo

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>

* Add graph to TensorBoard

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>

* Resume from last saved state within epoch.

Previously the last saved state would only resume from the beginning of an epoch with saved network settings. The changes in this code allow to resume from the last saved state within the epoc at the batch level. This should prevent overfitting by biasing the first sets of batches within an epoch that is saved and resumed.

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>

* Remove redundant line

Signed-off-by: Norman Heckscher <norman.heckscher@gmail.com>