Update for RNNCell
Changes to Tensorflow at the head neccesitate the need to create new instances of the RNNCell if using a different set of weights. In May 2017, Tensorflow will start transitioning this cell's behavior to use existing stored weights, if any, when it is called with scope=None (which can lead to silent model degradation).